Move common types to common file
