+++ pkg/ebpf/verifier/stats_no_linux.go
@@ -14,36 +14,6 @@ import (
 	"regexp"
 )
 
-// Statistics represent that statistics exposed via
-// the eBPF verifier when  LogLevelStats is enabled
-type Statistics struct {
-	StackDepth                 int `json:"stack_usage"`
-	InstructionsProcessed      int `json:"instruction_processed"`
-	InstructionsProcessedLimit int `json:"limit"`
-	MaxStatesPerInstruction    int `json:"max_states_per_insn"`
-	TotalStates                int `json:"total_states"`
-	PeakStates                 int `json:"peak_states"`
-}
-
-// ComplexityInfo holds the results of the complexity parsing
-type ComplexityInfo struct {
-}
-
-// StatsResult holds the result of the verifier stats process
-type StatsResult struct {
-	Stats           map[string]*Statistics
-	Complexity      map[string]*ComplexityInfo
-	FuncsPerSection map[string]map[string][]string
-}
-
-// StatsOptions holds the options for the function BuildVerifierStats
-type StatsOptions struct {
-	ObjectFiles        []string
-	FilterPrograms     []*regexp.Regexp
-	DetailedComplexity bool
-	VerifierLogsDir    string
-}
-
 // BuildVerifierStats accepts a list of eBPF object files and generates a
 // map of all programs and their Statistics
 func BuildVerifierStats(_ *StatsOptions) (*StatsResult, map[string]struct{}, error) {
